adding `page` options to the snippet
Adding `.page` options to the snippet. This commit also minifies at
runtime because handlebars + the minifier don't play well. If it's a
large enough perf hit then we should move to make the snippet code async
or at the very least memoized. Quick benchmark says `3-4ms` to minify,
so may be worth making it async soon.