Pin variable name for global key
Previous release of snippet introduced a new variable named `globalAnalyticsKey`, which after
getting mangled by minifier became `i`. Unfortunately, all the documentation now refers to this
variable as `i`. It is possible that the when next version of snippet is released, the same variable
gets a different mangled name, say `b`. This means that the documentation will be voided. To prevent
this, we're now locking the variable name to `i` forever. We can't change it now since the documentation
is already in the wild.